Remove z_stream from the Inflater and Deflater protocols
The underlying z_stream must be suitably encapsulated within the
Inflater/Deflater and need not be exposed to the users of the
latter. To enable closing this underlying z_stream, we could have
end() methods on the Inflater/Deflater.

This commit also includes some documentation improvements and a
warning is also fixed.